restructures component file structure.
