+++ src/Components/App/App.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import './App.css';
+import LoginForm from '../LoginForm/LoginForm';
+
+function App() {
+  return (
+    <div className="App">
+      <LoginForm />
+    </div>
+  );
+}
+
+export default App;
