+++ src/index.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
-import App from './App';
+import App from '../src/Components/App/App'
 import reportWebVitals from './reportWebVitals';
 
 const root = ReactDOM.createRoot(
